+++ plugin-locationcomponent/src/test/java/com/mapbox/maps/plugin/locationcomponent/DefaultLocationProviderTest.kt
@@ -417,7 +417,7 @@ private open class EmptyLocationConsumer : LocationConsumer {
   override fun onError(error: LocationError) {
   }
 
-  override fun onAccuracyRadiusUpdated(
+  override fun onHorizontalAccuracyRadiusUpdated(
     vararg radius: Double,
     options: (ValueAnimator.() -> Unit)?
   ) {